# Dashboard/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sharecycle API",
    "description": "API documentation for the Sharecycle backend application.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/bicycles/{id}": {
      "get": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "getBicycleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BicycleResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "updateBicycleMultipartAtId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "availableFrom",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BicycleResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "deleteBicycle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bicycles/admin/{id}": {
      "put": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "updateBicycleAsAdminJson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Bicycle"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bicycles": {
      "get": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "getAllBicycles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BicycleResponseDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "addBicycle",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "availableFrom",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BicycleResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/bicycles/order/preview": {
      "post": {
        "tags": [
          "bicycle-controller"
        ],
        "operationId": "previewOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "iphoneNumber": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "idCardNumber": {
            "type": "string"
          },
          "idCardImageUrl": {
            "type": "string"
          }
        }
      },
...
      "OrderRequestDTO": {
        "type": "object",
        "properties": {
          "bicycleId": {
            "type": "integer",
            "format": "int64"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          }
        }
      },
      "OrderResponseDTO": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "bicycleId": {
            "type": "integer",
            "format": "int64"
          },
          "bikeName": {
            "type": "string"
          },
          "bikeImageUrl": {
            "type": "string"
          },
          "pricePerDay": {
            "type": "number",
            "format": "double"
          },
          "renter": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "owner": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          },
          "rentalDays": {
            "type": "integer",
            "format": "int64"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "orderStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELED"
            ]
          }
        }
      },
      "BicycleResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availableFrom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "ownerEmail": {
            "type": "string"
          }
        }
      },
      "BicycleDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availableFrom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Bicycle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availableFrom": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "role": {
            "type": "string"
          },
          "iphoneNumber": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "idCardNumber": {
            "type": "string"
          },
          "idCardImageUrl": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          },
          "otpExpiryTime": {
            "type": "string",
            "format": "date-time"
          },
          "otpAttemptCount": {
            "type": "integer",
            "format": "int32"
          },
          "fcmToken": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}